refactor(user): rename the package from user to member
