+++ src/main/kotlin/apply/application/CheaterService.kt
@@ -3,9 +3,9 @@ package apply.application
 import apply.domain.cheater.Cheater
 import apply.domain.cheater.CheaterRepository
 import apply.domain.cheater.getOrThrow
-import apply.domain.user.MemberRepository
-import apply.domain.user.findAllByEmailIn
-import apply.domain.user.findByEmail
+import apply.domain.member.MemberRepository
+import apply.domain.member.findAllByEmailIn
+import apply.domain.member.findByEmail
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
