+++ src/main/kotlin/apply/application/UserAuthenticationService.kt
@@ -3,10 +3,10 @@ package apply.application
 import apply.domain.authenticationcode.AuthenticationCode
 import apply.domain.authenticationcode.AuthenticationCodeRepository
 import apply.domain.authenticationcode.getLastByEmail
-import apply.domain.user.UnidentifiedMemberException
-import apply.domain.user.MemberRepository
-import apply.domain.user.existsByEmail
-import apply.domain.user.findByEmail
+import apply.domain.member.UnidentifiedMemberException
+import apply.domain.member.MemberRepository
+import apply.domain.member.existsByEmail
+import apply.domain.member.findByEmail
 import apply.security.JwtTokenProvider
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
