+++ src/main/kotlin/apply/application/mail/MailService.kt
@@ -6,9 +6,9 @@ import apply.domain.mail.MailHistory
 import apply.domain.mail.MailHistoryRepository
 import apply.domain.recruitment.RecruitmentRepository
 import apply.domain.recruitment.getOrThrow
-import apply.domain.user.PasswordResetEvent
-import apply.domain.user.MemberRepository
-import apply.domain.user.getOrThrow
+import apply.domain.member.PasswordResetEvent
+import apply.domain.member.MemberRepository
+import apply.domain.member.getOrThrow
 import org.springframework.boot.autoconfigure.mail.MailProperties
 import org.springframework.core.io.ByteArrayResource
 import org.springframework.scheduling.annotation.Async
