+++ src/main/kotlin/apply/domain/member/Gender.kt
@@ -1,4 +1,4 @@
-package apply.domain.user
+package apply.domain.member
 
 enum class Gender(val title: String) {
     MALE("남성"),
