+++ src/main/kotlin/apply/domain/member/MemberInformation.kt
@@ -1,4 +1,4 @@
-package apply.domain.user
+package apply.domain.member
 
 import java.time.LocalDate
 import javax.persistence.Column
