+++ src/main/kotlin/apply/domain/member/MemberRepository.kt
@@ -1,4 +1,4 @@
-package apply.domain.user
+package apply.domain.member
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
