+++ src/main/kotlin/apply/domain/member/Password.kt
@@ -1,4 +1,4 @@
-package apply.domain.user
+package apply.domain.member
 
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParser
