+++ src/main/kotlin/apply/domain/member/UnidentifiedMemberException.kt
@@ -1,3 +1,3 @@
-package apply.domain.user
+package apply.domain.member
 
 class UnidentifiedMemberException(message: String? = null) : RuntimeException(message)
