+++ src/main/kotlin/apply/ui/api/AdministratorRestController.kt
@@ -3,7 +3,7 @@ package apply.ui.api
 import apply.application.AdministratorData
 import apply.application.AdministratorResponse
 import apply.application.AdministratorService
-import apply.domain.user.Member
+import apply.domain.member.Member
 import apply.security.LoginUser
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
