+++ src/main/kotlin/apply/ui/api/CheaterRestController.kt
@@ -3,7 +3,7 @@ package apply.ui.api
 import apply.application.CheaterData
 import apply.application.CheaterResponse
 import apply.application.CheaterService
-import apply.domain.user.Member
+import apply.domain.member.Member
 import apply.security.LoginUser
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
