+++ src/main/kotlin/apply/ui/api/TermRestController.kt
@@ -3,7 +3,7 @@ package apply.ui.api
 import apply.application.TermData
 import apply.application.TermResponse
 import apply.application.TermService
-import apply.domain.user.Member
+import apply.domain.member.Member
 import apply.security.LoginUser
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
