+++ src/main/kotlin/apply/ui/api/UserRestController.kt
@@ -9,7 +9,7 @@ import apply.application.UserAuthenticationService
 import apply.application.UserResponse
 import apply.application.UserService
 import apply.application.mail.MailService
-import apply.domain.user.Member
+import apply.domain.member.Member
 import apply.security.LoginUser
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
