+++ src/test/kotlin/apply/application/CheaterServiceTest.kt
@@ -3,8 +3,8 @@ package apply.application
 import apply.createCheater
 import apply.createCheaterData
 import apply.domain.cheater.CheaterRepository
-import apply.domain.user.MemberRepository
-import apply.domain.user.findByEmail
+import apply.domain.member.MemberRepository
+import apply.domain.member.findByEmail
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.nulls.shouldBeNull
