+++ src/test/kotlin/apply/application/EvaluationTargetIntegrationTest.kt
@@ -27,8 +27,8 @@ import apply.domain.evaluationtarget.EvaluationStatus.WAITING
 import apply.domain.evaluationtarget.EvaluationTarget
 import apply.domain.evaluationtarget.EvaluationTargetRepository
 import apply.domain.evaluationtarget.getOrThrow
-import apply.domain.user.Member
-import apply.domain.user.MemberRepository
+import apply.domain.member.Member
+import apply.domain.member.MemberRepository
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.extensions.spring.SpringTestExtension
 import io.kotest.extensions.spring.SpringTestLifecycleMode
