+++ src/test/kotlin/apply/application/UserAuthenticationServiceTest.kt
@@ -12,10 +12,10 @@ import apply.createRegisterUserRequest
 import apply.createUser
 import apply.domain.authenticationcode.AuthenticationCodeRepository
 import apply.domain.authenticationcode.getLastByEmail
-import apply.domain.user.UnidentifiedMemberException
-import apply.domain.user.MemberRepository
-import apply.domain.user.existsByEmail
-import apply.domain.user.findByEmail
+import apply.domain.member.UnidentifiedMemberException
+import apply.domain.member.MemberRepository
+import apply.domain.member.existsByEmail
+import apply.domain.member.findByEmail
 import apply.security.JwtTokenProvider
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
